imported freebsd53

+++ freebsd5/sys/dev/advansys/advmcode.h
@@ -1,7 +1,7 @@
 /*
  * Exported interface to downloadable microcode for AdvanSys SCSI Adapters
  *
- * $FreeBSD: src/sys/dev/advansys/advmcode.h,v 1.6 2000/01/14 03:33:38 gibbs Exp $
+ * $FreeBSD: src/sys/dev/advansys/advmcode.h,v 1.7 2004/03/10 20:52:47 phk Exp $
  *
  * Obtained from:
  *
@@ -14,6 +14,6 @@
  * modification.
  */
 
-extern u_int16_t adv_mcode[];
+extern u_int8_t adv_mcode[];
 extern u_int16_t adv_mcode_size;
 extern u_int32_t adv_mcode_chksum;
